refacotered the tcp send code
