+++ StreamDeckPlugin/Actions/DisplayCardAction.cs
@@ -1,11 +1,8 @@
-﻿using System.Net;
-using System.Net.Sockets;
-using System.Text;
+﻿using System.Text;
 using System.Threading.Tasks;
 using ArkhamOverlay.TcpUtils;
 using ArkhamOverlay.TcpUtils.Requests;
 using ArkhamOverlay.TcpUtils.Responses;
-using Newtonsoft.Json;
 using SharpDeck;
 using SharpDeck.Events.Received;
 using SharpDeck.Manifest;
@@ -41,28 +38,9 @@ namespace ArkhamOverlaySdPlugin.Actions {
 
         protected override Task OnWillAppear(ActionEventArgs<AppearancePayload> args) {
             var cardIndex = GetCardIndex(args.Payload.Coordinates);
-
-            var ipHostInfo = Dns.GetHostEntry(Dns.GetHostName());
-            var ipAddress = ipHostInfo.AddressList[0];
-            var remoteEP = new IPEndPoint(ipAddress, TcpInfo.Port);
-
-            var sender = new Socket(ipAddress.AddressFamily, SocketType.Stream, ProtocolType.Tcp);
-
             try {
-                sender.Connect(remoteEP);
-
                 var request = new GetCardInfoRequest { Deck = "Player1", Index = cardIndex };
-                var payload = Encoding.ASCII.GetBytes(request.ToString());
-
-                int bytesSent = sender.Send(payload);
-
-                var bytes = new byte[1024];
-                int bytesRec = sender.Receive(bytes);
-                var responseData = Encoding.ASCII.GetString(bytes, 0, bytesRec);
-                var response = JsonConvert.DeserializeObject<CardInfoReponse>(responseData.Substring(0, responseData.IndexOf("<EOF>")));
-
-                sender.Shutdown(SocketShutdown.Both);
-                sender.Close();
+                var response = SendSocketService.SendRequest<CardInfoReponse>(request);
 
                 return SetTitleAsync(WrapTitle(response.Name));
             } catch {
@@ -72,28 +50,9 @@ namespace ArkhamOverlaySdPlugin.Actions {
 
         protected override Task OnKeyDown(ActionEventArgs<KeyPayload> args) {
             var cardIndex = GetCardIndex(args.Payload.Coordinates);
-
-            var ipHostInfo = Dns.GetHostEntry(Dns.GetHostName());
-            var ipAddress = ipHostInfo.AddressList[0];
-            var remoteEP = new IPEndPoint(ipAddress, TcpInfo.Port);
-
-            var sender = new Socket(ipAddress.AddressFamily, SocketType.Stream, ProtocolType.Tcp);
-
             try {
-                sender.Connect(remoteEP);
-
                 var request = new ClickCardButtonRequest { Deck = "Player1", Index = cardIndex };
-                var payload = Encoding.ASCII.GetBytes(request.ToString());
-
-                int bytesSent = sender.Send(payload);
-
-                var bytes = new byte[1024];
-                int bytesRec = sender.Receive(bytes);
-                var responseData = Encoding.ASCII.GetString(bytes, 0, bytesRec);
-                var response = JsonConvert.DeserializeObject<CardInfoReponse>(responseData.Substring(0, responseData.IndexOf("<EOF>")));
-
-                sender.Shutdown(SocketShutdown.Both);
-                sender.Close();
+                var response = SendSocketService.SendRequest<CardInfoReponse>(request);
 
                 return SetTitleAsync(WrapTitle(response.Name));
             } catch {
